Use uint64_t timestamp to overcome potential overflow
Added explicit casting to uint64_t to overcome potential overflow
within some expressions and modified the containing type of a timestamp
variable to store all bits of a time_t value

+++ lld/COFF/Config.h
@@ -290,7 +290,7 @@ struct Configuration {
   uint32_t minorOSVersion = 0;
   uint32_t majorSubsystemVersion = 6;
   uint32_t minorSubsystemVersion = 0;
-  uint32_t timestamp = 0;
+  uint64_t timestamp = 0;
   uint32_t functionPadMin = 0;
   uint32_t timeTraceGranularity = 0;
   uint16_t dependentLoadFlags = 0;
